requirements.txt
streamlit==1.44.1
openai==1.71.0
python-dotenv==1.1.0
streamlit-option-menu==0.4.0